requirements.txt
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
bayesian-optimization==1.3.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cma==3.2.2
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
feedparser==6.0.10
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
nevergrad==0.5.0
notebook==6.4.12
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pycparser==2.21
pydeck==0.8.0b4
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.1
seaborn==0.12.0
semver==2.13.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
srt==3.5.2
stack-data==0.5.1
streamlit==1.13.0
tenacity==8.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
vosk==0.3.44
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
widgetsnbextension==4.0.3
zipp==3.8.1
